Fix error checking in packet handlers, increase min. packet lengths

+++ include/server/server/packet.h
@@ -94,10 +94,10 @@ extern void packet_free(bedrock_packet *packet);
 
 extern int packet_parse(struct client *client, const bedrock_packet *packet);
 
-extern void packet_read(const bedrock_packet *packet, size_t *offset, void *dest, size_t dest_size);
-extern void packet_read_int(const bedrock_packet *packet, size_t *offset, void *dest, size_t dest_size);
-extern void packet_read_string(const bedrock_packet *packet, size_t *offset, char *dest, size_t dest_size);
-extern void packet_read_slot(const bedrock_packet *packet, size_t *offset, struct item_stack *stack);
+extern void packet_read(const bedrock_packet *packet, int *offset, void *dest, size_t dest_size);
+extern void packet_read_int(const bedrock_packet *packet, int *offset, void *dest, size_t dest_size);
+extern void packet_read_string(const bedrock_packet *packet, int *offset, char *dest, size_t dest_size);
+extern void packet_read_slot(const bedrock_packet *packet, int *offset, struct item_stack *stack);
 
 extern void packet_pack_header(bedrock_packet *packet, uint8_t header);
 extern void packet_pack(bedrock_packet *packet, const void *data, size_t size);
