+++ server/packet/packet_client_settings.c
@@ -3,7 +3,7 @@
 
 int packet_client_settings(struct client bedrock_attribute_unused *client, const bedrock_packet *packet)
 {
-	size_t offset = PACKET_HEADER_LENGTH;
+	int offset = PACKET_HEADER_LENGTH;
 	char locale[BEDROCK_MAX_STRING_LENGTH];
 	uint8_t view_distance, chat_flags, difficulty, cape;
